feat: add 500 error handling
